Relocate ParentSelectorWidget to Widgets directory
