+++ packages/airview-cms/src/features/cms/widgets/dynamic-widget.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { ParentSelectorWidget } from "../parent-selector-widget";
+import { ParentSelectorWidget } from "./parent-selector-widget";
 import { BooleanWidget } from "./boolean-widget";
 import { TextFieldWidget } from "./text-field-widget";
