+++ packages/airview-cms/src/features/cms/widgets/index.js
@@ -1,3 +1,4 @@
 export { DynamicWidget } from "./dynamic-widget";
 export { TextFieldWidget } from "./text-field-widget";
 export { BooleanWidget } from "./boolean-widget";
+export { ParentSelectorWidget } from "./parent-selector-widget";
